Add '/buffer showname' command as a11y workaround
For users with a refreshable braille display using brltty, WeeChat does
not display the current buffer name when in or switching to a buffer.
This makes it easy for someone dependent on such a device to lose their
bearings in the client. The '/buffer showname' command is designed to
work around this by getting the name of the current buffer and
displaying it in the current buffer whenever a user needs to re-gain
their bearings.

Although the command '/buffer get name' already exists, this prints the
name to the main weechat buffer which isn't as useful and is cumbersome
if one has to switch to the main weechat buffer, read the most recent
text, then switch back. Furthermore, it also prints text which looks
something like:

irc.server.freenode: (str) short_name = freenode

which contains extra unhelpful information for someone just trying to
find out which channel they are in.